core-bankec-python/requirements.txt
Flask==2.0.3
psycopg2-binary==2.9.10
gunicorn==20.1.0
python-dotenv
Flask-RESTX==0.5.1
Werkzeug==2.0.3
PyJWT==2.8.0
bcrypt==4.0.1
bleach==6.0.0
